+++ .github/workflows/argilla.yml
@@ -21,7 +21,7 @@ jobs:
   build:
     services:
       argilla-server:
-        image: argilladev/argilla-hf-spaces:develop
+        image: argilladev/argilla-hf-spaces:pr-5573
         ports:
           - 6900:6900
         env:
