+++ argilla-server/src/argilla_server/api/schemas/v1/fields.py
@@ -21,7 +21,6 @@ from argilla_server.enums import FieldType
 from argilla_server.pydantic_v1 import BaseModel, constr
 from argilla_server.pydantic_v1 import Field as PydanticField
 
-FIELD_CREATE_NAME_REGEX = r"^(?=.*[a-z0-9])[a-z0-9_-]+$"
 FIELD_CREATE_NAME_MIN_LENGTH = 1
 FIELD_CREATE_NAME_MAX_LENGTH = 200
 
@@ -30,7 +29,6 @@ FIELD_CREATE_TITLE_MAX_LENGTH = 500
 
 FieldName = Annotated[
     constr(
-        regex=FIELD_CREATE_NAME_REGEX,
         min_length=FIELD_CREATE_NAME_MIN_LENGTH,
         max_length=FIELD_CREATE_NAME_MAX_LENGTH,
     ),
