+++ argilla-server/tests/unit/api/handlers/v1/datasets/test_questions.py
@@ -329,24 +329,6 @@ class TestDatasetQuestions:
         assert response.status_code == 403
         assert (await db.execute(select(func.count(Question.id)))).scalar() == 0
 
-    @pytest.mark.parametrize("invalid_name", ["", " ", "  ", "-", "--", "_", "__", "A", "AA", "invalid_nAmE"])
-    async def test_create_dataset_question_with_invalid_name(
-        self, async_client: "AsyncClient", db: "AsyncSession", owner_auth_header: dict, invalid_name: str
-    ):
-        dataset = await DatasetFactory.create()
-        question_json = {
-            "name": invalid_name,
-            "title": "title",
-            "settings": {"type": "text"},
-        }
-
-        response = await async_client.post(
-            f"/api/v1/datasets/{dataset.id}/questions", headers=owner_auth_header, json=question_json
-        )
-
-        assert response.status_code == 422
-        assert (await db.execute(select(func.count(Question.id)))).scalar() == 0
-
     async def test_create_dataset_question_with_invalid_max_length_name(
         self, async_client: "AsyncClient", db: "AsyncSession", owner_auth_header: dict
     ):
