+++ argilla/src/argilla/_models/_settings/_fields.py
@@ -56,12 +56,6 @@ class FieldModel(ResourceModel):
     description: Optional[str] = None
     dataset_id: Optional[UUID] = None
 
-    @field_validator("name")
-    @classmethod
-    def __name_lower(cls, name):
-        formatted_name = name.lower().replace(" ", "_")
-        return formatted_name
-
     @field_validator("title")
     @classmethod
     def __title_default(cls, title: str, info: ValidationInfo) -> str:
