+++ argilla/src/argilla/settings/_io/_hub.py
@@ -220,12 +220,6 @@ def _define_settings_from_features(
         feature_type = _map_feature_type(feature)
         attribute_definition = _map_attribute_type(feature_mapping.get(name))
 
-        name = Settings._sanitize_settings_name(name)
-
-        if not Settings._is_valid_name(name):
-            warnings.warn(f"Feature '{name}' has an invalid name. Skipping.")
-            continue
-
         if feature_type == FeatureType.CHAT:
             fields.append(ChatField(name=name, required=False))
         elif feature_type == FeatureType.TEXT:
